Plotting+Bliss+HSA implemented
In Combinations I implemented get_bliss_response, added _check_bliss_requirements, implemented get_hsa_response and _check_drug_consistency. Problem: When running the class the test code at the bottom of the class is used twice for an unknown reason.
In class Drug I deleted the inputs control_response and monotone_increasing of get_reponse as we can use the values of the Drug. I also added if statements in get_response to check that we do not try to calculate log(0) as this threw errors.
I implemented Plotting.